Change how error checking is done.
Old version was:

if ($response->ack ==== 'Failure') {
    // Show error messages
} else {
    // Display results.
}

The problem is that the code assumes that error messages are only going
to be present if ack is equal to Failure. In some situations there could
be error messages and the value of ack is in fact Warning. It is also
possible that ack is never returned by the service. The new code shown
below should be better at handling these types of situations.

if ($response->ack !== 'Success') {
    // Show error messages
} else {
    // Display results.
}